@@ -1,16 +1,11 @@
-import sys
 import uuid
+from collections.abc import AsyncGenerator, Awaitable, Callable
 from logging import getLogger
 from typing import (
     TYPE_CHECKING,
     Any,
-    AsyncGenerator,
-    Awaitable,
-    Callable,
-    Dict,
-    Optional,
+    TypeAlias,
     TypeVar,
-    Union,
 )
 
 from redis.asyncio import BlockingConnectionPool, Connection, Redis, ResponseError
@@ -23,10 +18,6 @@ _T = TypeVar("_T")
 
 logger = getLogger("taskiq.redis_broker")
 
-if sys.version_info >= (3, 10):
-    from typing import TypeAlias
-else:
-    from typing_extensions import TypeAlias
 
 if TYPE_CHECKING:
     _BlockingConnectionPool: TypeAlias = BlockingConnectionPool[Connection]  # type: ignore
@@ -40,10 +31,10 @@ class BaseRedisBroker(AsyncBroker):
     def __init__(
         self,
         url: str,
-        task_id_generator: Optional[Callable[[], str]] = None,
-        result_backend: Optional[AsyncResultBackend[_T]] = None,
+        task_id_generator: Callable[[], str] | None = None,
+        result_backend: AsyncResultBackend[_T] | None = None,
         queue_name: str = "taskiq",
-        max_connection_pool_size: Optional[int] = None,
+        max_connection_pool_size: int | None = None,
         **connection_kwargs: Any,
     ) -> None:
         """
@@ -160,18 +151,18 @@ class RedisStreamBroker(BaseRedisBroker):
         self,
         url: str,
         queue_name: str = "taskiq",
-        max_connection_pool_size: Optional[int] = None,
+        max_connection_pool_size: int | None = None,
         consumer_group_name: str = "taskiq",
-        consumer_name: Optional[str] = None,
+        consumer_name: str | None = None,
         consumer_id: str = "$",
         mkstream: bool = True,
         xread_block: int = 2000,
-        maxlen: Optional[int] = None,
+        maxlen: int | None = None,
         approximate: bool = True,
         idle_timeout: int = 600000,  # 10 minutes
         unacknowledged_batch_size: int = 100,
-        xread_count: Optional[int] = 100,
-        additional_streams: Optional[Dict[str, Union[str, int]]] = None,
+        xread_count: int | None = 100,
+        additional_streams: dict[str, str | int] | None = None,
         **connection_kwargs: Any,
     ) -> None:
         """
